modified sample data
