+++ src/test/java/seedu/address/testutil/PersonBuilder.java
@@ -12,7 +12,7 @@ import seedu.address.model.participant.Participant;
 import seedu.address.model.participant.Phone;
 import seedu.address.model.participant.Sex;
 import seedu.address.model.tag.Tag;
-import seedu.address.model.util.SampleParticipantsUtil;
+import seedu.address.model.util.SampleDataUtil;
 
 /**
  * A utility class to help with building Participant objects.
@@ -90,7 +90,7 @@ public class PersonBuilder {
      * Parses the {@code tags} into a {@code Set<Tag>} and set it to the {@code Participant} that we are building.
      */
     public PersonBuilder withTags(String ... tags) {
-        this.tags = SampleParticipantsUtil.getTagSet(tags);
+        this.tags = SampleDataUtil.getTagSet(tags);
         return this;
     }
